+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+* `flexberry-lookup` component:
+    * Dropdown mode uses projection from `projection` property.
 
 ## [2.1.0-beta.3] - 2019-03-15
 ### Added